fix: change age_minutes type from DOUBLE PRECISION to NUMERIC in scraper health check function

+++ supabase/migrations/20260215600000_scraper_health_check.sql
@@ -11,7 +11,7 @@ CREATE OR REPLACE FUNCTION public.check_scraper_health(
 )
 RETURNS TABLE (
   last_observed_at TIMESTAMPTZ,
-  age_minutes DOUBLE PRECISION,
+  age_minutes NUMERIC,
   is_healthy BOOLEAN
 )
 LANGUAGE plpgsql
